requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
httpx==0.27.0
py3xui==0.2.1
python-dotenv==1.0.1
